+++ test/engine/engine_sensor_test.cc
@@ -1753,6 +1753,8 @@ TEST_F(SensorTest, InsideSiteFlexBody) {
   )";
 
   char error[1024] = {0};
+  EXPECT_CALL(mock_warning_handler, Warn(testing::HasSubstr("is not rigid")))
+      .WillOnce(testing::Return());
   mjModel* m = LoadModelFromString(xml, error, sizeof(error));
   ASSERT_THAT(m, NotNull()) << error;
   mjData* d = mj_makeData(m);
@@ -1839,6 +1841,8 @@ TEST_F(SensorTest, FlexContactSensors) {
   )";
 
   char error[1024] = {0};
+  EXPECT_CALL(mock_warning_handler, Warn(testing::HasSubstr("is not rigid")))
+      .WillOnce(testing::Return());
   mjModel* m = LoadModelFromString(xml, error, sizeof(error));
   ASSERT_THAT(m, NotNull()) << error;
   mjData* d = mj_makeData(m);